requirements.txt
gunicorn==19.8.1
sendgrid==5.4.0
celery-redbeat==0.11.1
celery==4.2.0
click==6.7
Flask==0.12.1
Flask-Script==2.0.6
flask-migrate==2.2.0
flask-sqlalchemy==2.3.2
flask-migrate==2.2.0
pymysql==0.8.1
redis==2.10.6
requests==2.19.1
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.9
